Move colors to modificationData.js
